init TreeBracketFor 16
