+++ backend/src/test/java/com/rikishi/rikishi/system/TreeBracketTest.java
@@ -1,0 +1,11 @@
+package com.rikishi.rikishi.system;
+
+import org.junit.jupiter.api.Test;
+
+
+public class TreeBracketTest {
+    @Test
+    void genrateDuelsFor8Test() {
+
+    }
+}
